intern/backend/requirements.txt
streamlit
transformers
torch
langchain-community
langchain-core
pandas==1.5.3
streamlit==1.22.0
transformers==4.25.1
torch==1.13.1
python-dotenv==0.21.0
langchain_community==0.0.6
langchain_core==0.0.166
chroma-vector-store==0.1.0
